refactor groovy
