+++ lib/manager/gradle2/extract.js
@@ -1,12 +1,20 @@
-const { parseBuildGradle } = require('./parse');
+const { parseBuildGradle } = require('./parse/groovy');
 
 module.exports = {
   extractPackageFile,
 };
 
-function extractPackageFile(content) {
+function extractPackageFile(content, filename) {
+  let gradleFileType;
   let deps = [];
-  const parsed = parseBuildGradle(content);
+  let parsed;
+  if (filename && filename.endsWith('build.gradle')) {
+    gradleFileType = 'groovy';
+    parsed = parseBuildGradle(content);
+  }
+  if (!parsed) {
+    return null;
+  }
   logger.debug({ parsed });
   if (parsed.plugins) {
     deps = parsed.plugins.map(plugin => {
@@ -30,5 +38,5 @@ function extractPackageFile(content) {
   if (!deps.length) {
     return null;
   }
-  return { deps };
+  return { gradleFileType, deps };
 }
