+++ lib/manager/gradle2/update/groovy.js
@@ -1,4 +1,4 @@
-const { pluginRegex } = require('./parse');
+const { pluginRegex } = require('../parse/groovy');
 
 module.exports = {
   updateDependency,
